requirements.txt
streamlit>=1.22
google-generativeai>=0.3.0
pandas>=1.5.0
pdfplumber>=0.7.0
reportlab>=4.0.0
python-dotenv>=0.21.0